add more tests.

+++ run_and_clean.sh
@@ -2,5 +2,16 @@
 
 ./../valhalla/build/*/images/jdk/bin/java -cp library/target/scala-library-nonbootstrapped/scala-library-3.8.1-RC1-bin-SNAPSHOT-nonbootstrapped.jar:tests/valhalla/bin --enable-preview Main
 
-cd tests/valhalla/bin
-rm -rf *
+clean="true"
+
+while getopts n flag
+do
+    case "${flag}" in
+        n) clean="false";;
+    esac
+done
+
+if [[ "$clean" == "true" ]]; then
+  cd tests/valhalla/bin
+  rm -rf *
+fi
